fix: narrow express-button gate to product purchasability
Per review (#11881), the PR bundled two changes: hiding express
buttons when the product isn't purchasable/in stock (what WOOPMNT-4231
asks for), and hiding them on variable products until a valid variation
is selected (a UX change beyond the issue). Revert the second: restore
the prior "always show on variations, prompt on click" behavior and the
server-rendered placeholder. Keep only the server-side
is_product_purchasable() gate wired into both should_show_* methods.
