+++ client/express-checkout/shortcode-buttons-express/__tests__/tokenized-express-checkout--product-page.test.js
@@ -652,68 +652,4 @@ describe( 'Tokenized Express Checkout Element - Product page logic', () => {
 			} )
 		);
 	} );
-
-	it( 'hides the express button on ready when add to cart is blocked', async () => {
-		document
-			.querySelector( '.single_add_to_cart_button' )
-			.classList.add( 'disabled' );
-
-		await jest.isolateModulesAsync( async () => {
-			await import( '..' );
-		} );
-
-		// triggering the `ready` event with available methods.
-		stripeElementMock.__getRegisteredEvent( 'ready' )( {
-			availablePaymentMethods: {
-				link: false,
-				applePay: true,
-				googlePay: true,
-				paypal: false,
-				amazonPay: false,
-			},
-		} );
-
-		expect(
-			screen.getByTestId( 'wcpay-express-checkout-element' )
-		).not.toBeVisible();
-	} );
-
-	it( 'hides the express button when the selection becomes blocked', async () => {
-		apiFetch.mockImplementation( async () =>
-			Promise.resolve( {
-				json: () => Promise.resolve( { items: [] } ),
-				headers: new Map(),
-			} )
-		);
-
-		let doAction;
-		await jest.isolateModulesAsync( async () => {
-			( { doAction } = await import( '@wordpress/hooks' ) );
-			await import( '..' );
-		} );
-
-		// becomes visible on ready with available methods.
-		stripeElementMock.__getRegisteredEvent( 'ready' )( {
-			availablePaymentMethods: {
-				link: false,
-				applePay: true,
-				googlePay: true,
-				paypal: false,
-				amazonPay: false,
-			},
-		} );
-		expect(
-			screen.getByTestId( 'wcpay-express-checkout-element' )
-		).toBeVisible();
-
-		// shopper picks an unavailable variation: add-to-cart becomes disabled.
-		document
-			.querySelector( '.single_add_to_cart_button' )
-			.classList.add( 'disabled' );
-		doAction( 'wcpay.express-checkout.update-button-data' );
-
-		expect(
-			screen.getByTestId( 'wcpay-express-checkout-element' )
-		).not.toBeVisible();
-	} );
 } );
